+++ build-src/after_install.sh
@@ -4,14 +4,17 @@
 systemctl stop tuxedofancontrol || true
 systemctl disable tuxedofancontrol || true
 
+DIST_DATA=/opt/tuxedo-control-center/resources/dist/tuxedo-control-center/data/dist-data
+
 rm /usr/share/applications/tuxedo-control-center.desktop || true
-cp /opt/tuxedo-control-center/resources/dist/tuxedo-control-center/data/dist-data/tuxedo-control-center.desktop /usr/share/applications/tuxedo-control-center.desktop || true
-cp /opt/tuxedo-control-center/resources/dist/tuxedo-control-center/data/dist-data/de.tuxedocomputers.tcc.policy /usr/share/polkit-1/actions/de.tuxedocomputers.tcc.policy || true
-cp /opt/tuxedo-control-center/resources/dist/tuxedo-control-center/data/dist-data/com.tuxedocomputers.tccd.conf /usr/share/dbus-1/system.d/com.tuxedocomputers.tccd.conf || true
+cp ${DIST_DATA}/tuxedo-control-center.desktop /usr/share/applications/tuxedo-control-center.desktop || true
+cp ${DIST_DATA}/tuxedo-control-center-tray.desktop /etc/skel/.config/autostart/tuxedo-control-center-tray.desktop || true
+cp ${DIST_DATA}/de.tuxedocomputers.tcc.policy /usr/share/polkit-1/actions/de.tuxedocomputers.tcc.policy || true
+cp ${DIST_DATA}/com.tuxedocomputers.tccd.conf /usr/share/dbus-1/system.d/com.tuxedocomputers.tccd.conf || true
 
 # Copy and enable services
-cp /opt/tuxedo-control-center/resources/dist/tuxedo-control-center/data/dist-data/tccd.service /etc/systemd/system/tccd.service || true
-cp /opt/tuxedo-control-center/resources/dist/tuxedo-control-center/data/dist-data/tccd-sleep.service /etc/systemd/system/tccd-sleep.service || true
+cp ${DIST_DATA}/tccd.service /etc/systemd/system/tccd.service || true
+cp ${DIST_DATA}/tccd-sleep.service /etc/systemd/system/tccd-sleep.service || true
 systemctl daemon-reload
 systemctl enable tccd tccd-sleep
 systemctl start tccd