+++ build-src/electron-builder.ts
@@ -50,6 +50,7 @@ async function buildDeb(): Promise<void> {
             distSrc + '/data/dist-data/tccd-sleep.service',
             distSrc + '/data/dist-data/tuxedo-control-center_256.svg',
             distSrc + '/data/dist-data/tuxedo-control-center.desktop',
+            distSrc + '/data/dist-data/tuxedo-control-center-tray.desktop',
             distSrc + '/data/dist-data/de.tuxedocomputers.tcc.policy',
             distSrc + '/data/dist-data/com.tuxedocomputers.tccd.conf'
         ],
@@ -108,6 +109,7 @@ async function buildSuseRpm(): Promise<void> {
             distSrc + '/data/dist-data/tccd-sleep.service',
             distSrc + '/data/dist-data/tuxedo-control-center_256.svg',
             distSrc + '/data/dist-data/tuxedo-control-center.desktop',
+            distSrc + '/data/dist-data/tuxedo-control-center-tray.desktop',
             distSrc + '/data/dist-data/de.tuxedocomputers.tcc.policy',
             distSrc + '/data/dist-data/com.tuxedocomputers.tccd.conf'
         ],
@@ -166,6 +168,7 @@ async function buildAppImage(): Promise<void> {
             distSrc + '/data/dist-data/tuxedo-control-center_256.png',
             distSrc + '/data/dist-data/tuxedo-control-center_256.svg',
             distSrc + '/data/dist-data/tuxedo-control-center.desktop',
+            distSrc + '/data/dist-data/tuxedo-control-center-tray.desktop',
             distSrc + '/data/dist-data/de.tuxedocomputers.tcc.policy',
             distSrc + '/data/dist-data/com.tuxedocomputers.tccd.conf'
         ],