+++ src/e-app/main.ts
@@ -1,17 +1,121 @@
-import { app, BrowserWindow, ipcMain } from 'electron';
+import { app, BrowserWindow, ipcMain, Tray, Menu, globalShortcut } from 'electron';
 import * as path from 'path';
 import * as child_process from 'child_process';
+import * as fs from 'fs';
+import * as os from 'os';
 
-let win: Electron.BrowserWindow;
+// Tweak to get correct dirname for resource files outside app.asar
+const appPath = __dirname.replace('app.asar/', '');
 
-const watch = process.argv.includes('--watch');
+const autostartLocation = path.join(os.homedir(), '.config/autostart');
+const autostartDesktopFilename = 'tuxedo-control-center-tray.desktop';
+const tccConfigDir = '.tcc';
 
-if (watch) {
+let tccWindow: Electron.BrowserWindow;
+let tray: Electron.Tray;
+
+const watchOption = process.argv.includes('--watch');
+const trayOnlyOption = process.argv.includes('--tray');
+
+// Ensure that only one instance of the application is running
+const applicationLock = app.requestSingleInstanceLock();
+if (!applicationLock) {
+    console.log('TUXEDO Control Center is already running');
+    app.exit(0);
+}
+
+if (watchOption) {
     require('electron-reload')(path.join(__dirname, '..', 'ng-app'));
 }
 
-function createWindow() {
-    win = new BrowserWindow({
+if (isFirstStart()) {
+    installAutostartTray();
+}
+
+if (!userConfigDirExists()) {
+    createUserConfigDir();
+}
+
+app.on('second-instance', (event, cmdLine, workingDir) => {
+    // If triggered by a second instance, find/show/start GUI
+    activateTccGui();
+});
+
+app.whenReady().then( () => {
+    const success = globalShortcut.register('Super+Alt+F6', () => {
+        activateTccGui();
+    });
+    if (!success) { console.log('Failed to register global shortcut'); }
+
+    createTccTray();
+    if (!trayOnlyOption) {
+        activateTccGui();
+    }
+});
+
+app.on('will-quit', (event) => {
+    // Prevent default quit action
+    event.preventDefault();
+
+    // Close window but do not quit application unless tray is gone
+    if (tccWindow) {
+        tccWindow.close();
+        tccWindow = null;
+    }
+    if (!tray || tray.isDestroyed()) {
+        // Actually quit
+        globalShortcut.unregisterAll();
+        app.exit(0);
+    }
+});
+
+app.on('window-all-closed', () => {
+    if (!tray || tray.isDestroyed()) {
+        quitCurrentTccSession();
+    }
+});
+
+function activateTccGui() {
+    if (tccWindow) {
+        if (tccWindow.isMinimized()) { tccWindow.restore(); }
+        tccWindow.focus();
+    } else {
+        createTccWindow();
+    }
+}
+
+function createTccTray() {
+    const trayInstalled = isAutostartTrayInstalled();
+    const trayIcon =  path.join(__dirname, '../data/dist-data/tuxedo-control-center_256.png');
+    if (!tray) {
+        tray = new Tray(trayIcon);
+        tray.setTitle('TUXEDO Control Center');
+        tray.setToolTip('TUXEDO Control Center');
+    }
+    const contextMenu = Menu.buildFromTemplate([
+        { label: 'TUXEDO Control Center', type: 'normal', click: () => { activateTccGui(); } },
+        {
+                label: 'Tray autostart', type: 'checkbox', checked: trayInstalled,
+                click: () => trayInstalled ? menuRemoveAutostartTray() : menuInstallAutostartTray()
+        },
+        { type: 'separator' },
+        { label: 'Exit', type: 'normal', click: () => { quitCurrentTccSession(); } }
+    ]);
+    tray.setContextMenu(contextMenu);
+}
+
+function menuInstallAutostartTray() {
+    installAutostartTray();
+    createTccTray();
+}
+
+function menuRemoveAutostartTray() {
+    removeAutostartTray();
+    createTccTray();
+}
+
+function createTccWindow() {
+    tccWindow = new BrowserWindow({
         title: 'TUXEDO Control Center',
         width: 1040,
         height: 750,
@@ -19,32 +123,27 @@ function createWindow() {
         resizable: true,
         minWidth: 1040,
         minHeight: 750,
-        icon: path.join(__dirname, '../data/dist-data/tuxedo-control-center_256.svg'),
+        icon: path.join(__dirname, '../data/dist-data/tuxedo-control-center_256.png'),
         webPreferences: {
             nodeIntegration: true
         }
     });
 
     const indexPath = path.join(__dirname, '..', 'ng-app', 'index.html');
-    win.loadFile(indexPath);
-    win.on('closed', () => {
-        win = null;
+    tccWindow.loadFile(indexPath);
+    tccWindow.on('closed', () => {
+        tccWindow = null;
     });
 }
 
-app.on('ready', createWindow);
-
-app.on('window-all-closed', () => {
-    if (process.platform !== 'darwin') {
-        app.quit();
+function quitCurrentTccSession() {
+    if (tray) {
+        tray.destroy();
+        tray = null;
     }
-});
 
-app.on('activate', () => {
-    if (win === null) {
-        createWindow();
-    }
-});
+    app.quit();
+}
 
 ipcMain.on('exec-cmd-sync', (event, arg) => {
     try {
@@ -67,3 +166,58 @@ ipcMain.on('exec-cmd-async', (event, arg) => {
 ipcMain.on('spawn-external-async', (event, arg) => {
     child_process.spawn(arg, { detached: true, stdio: 'ignore' });
 });
+
+function installAutostartTray(): boolean {
+    try {
+        fs.copyFileSync(
+            path.join(appPath, '../data/dist-data', autostartDesktopFilename),
+            path.join(autostartLocation, autostartDesktopFilename)
+        );
+        return true;
+    } catch (err) {
+        console.log('Failed to install autostart tray -> ' + err);
+        return false;
+    }
+}
+
+function removeAutostartTray(): boolean {
+    try {
+        if (fs.existsSync(path.join(autostartLocation, autostartDesktopFilename))) {
+            fs.unlinkSync(path.join(autostartLocation, autostartDesktopFilename));
+        }
+        return true;
+    } catch (err) {
+        console.log('Failed to remove autostart tray -> ' + err);
+        return false;
+    }
+}
+
+function isAutostartTrayInstalled(): boolean {
+    try {
+        return fs.existsSync(path.join(autostartLocation, autostartDesktopFilename));
+    } catch (err) {
+        console.log('Failed to check if autostart tray is installed -> ' + err);
+        return false;
+    }
+}
+
+function isFirstStart(): boolean {
+    return !userConfigDirExists();
+}
+
+function userConfigDirExists(): boolean {
+    try {
+        return fs.existsSync(path.join(os.homedir(), '.tcc'));
+    } catch (err) {
+        return false;
+    }
+}
+
+function createUserConfigDir(): boolean {
+    try {
+        fs.mkdirSync(path.join(os.homedir(), '.tcc'));
+        return true;
+    } catch (err) {
+        return false;
+    }
+}