+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## [1.0.3] - Unreleased
 ### Added
 - Support for new devices
+- Basic tray icon
+- Global keyboard shortcut (Tux+Alt+F6) to start
 
 ### Changed
 - Revised fan tables for existing fan profiles
 - New fan profile *Silent*
+- New fan profile *Freezy*
 - Cool & Breezy now uses the *Quiet* fan profile
 - Powersave extreme now uses the *Silent* fan profile